+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added `From` implementations for various integer types
+  - Removed `Num::from_int` constructor
 
 
 0.1.0